add interpolation tests
